solving some errors

+++ src/main/java/com/secureChatWebApp/controllers/LoginController.java
@@ -25,7 +25,7 @@ public class LoginController {
 		Boolean isValidUser = (Boolean) request.getAttribute("isValidUser");
 
 		// get startTime added by Security intercepter
-		long startTime = (long) request.getAttribute("startTime");
+		Long startTime = (Long) request.getAttribute("startTime");
 
 		// response JSON
 		LinkedHashMap<String, String> responseJSON = new LinkedHashMap<String, String>();
@@ -33,7 +33,7 @@ public class LoginController {
 		// if user is not valid return error response JSON
 		// return a bad request status code (400)
 		if (!isValidUser.booleanValue()) {
-			double timeTaken = ((System.currentTimeMillis() - startTime) / 1000.0);
+			double timeTaken = ((System.currentTimeMillis() - startTime.longValue()) / 1000.0);
 			responseJSON.put("domain", "Login");
 			responseJSON.put("errMessage", "Invalid User. Register as a new User before login.");
 			responseJSON.put("timeTaken", timeTaken + " seconds");
@@ -43,7 +43,7 @@ public class LoginController {
 		// if user is not authenticated return error response message
 		// with UNAUTHORIZED status code (401)
 		if (!isAuthenticated.booleanValue()) {
-			double timeTaken = ((System.currentTimeMillis() - startTime) / 1000.0);
+			double timeTaken = ((System.currentTimeMillis() - startTime.longValue()) / 1000.0);
 			responseJSON.put("domain", "Login");
 			responseJSON.put("errMessage", "UnAuthorized user. Could not authenticate user. It may be because "
 					+ "request has delay over 4 minutes so please check your internet connection and check that it is secure to avoid reply attacks.");
@@ -53,7 +53,7 @@ public class LoginController {
 
 		// if user is authenticated and valid then send a success
 		// response Json
-		double timeTaken = ((System.currentTimeMillis() - startTime) / 1000.0);
+		double timeTaken = ((System.currentTimeMillis() - startTime.longValue()) / 1000.0);
 		responseJSON.put("domain", "Login");
 		responseJSON.put("message", "User logged in successfully");
 		responseJSON.put("timeTaken", timeTaken + " seconds");
